requirements.txt
python-telegram-bot==20.3
yt-dlp==2023.11.14
python-dotenv==1.0.0
